separete directory for optimizers
