+++ gsplat/optimizers/__init__.py
@@ -1,0 +1,1 @@
+from .selective_adam import SelectiveAdam
